requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
upstash-redis>=1.0.0
